Centralize GetVcountRandom prototype into include/game.h
True interface is s32 (body's (u8) cast emits the exit mask; callers use
result as int/u32); def return type was u8 — changed to s32, byte-neutral.
4 callers de-externed, proto inserted into game.h.

+++ src/engine/sub_0801185c.c
@@ -1,4 +1,5 @@
 #include "gba/io.h"
+#include "game.h"
 #include "iwram.h"
 #include "types.h"
 
@@ -6,7 +7,7 @@ extern int __modsi3(int num, int den);
 
 /* VCOUNT-seeded "pick a value in [0, range)" helper: scrambles the current
  * scanline with an LCG-style multiply-add, then folds it modulo `range`. */
-u8 GetVcountRandom(u8 range)
+s32 GetVcountRandom(u8 range)
 {
     return (u8)__modsi3((u8)REG_VCOUNT * 107 + 7, range);
 }
